+++ releasenotes.md
@@ -25,7 +25,7 @@
 - Allow specifying an import module using `@wasm` #1305.
 - Deprecated inline generic types outside of struct definitions and macros unless marked `@adhoc`.
 - Improved method detection in earlier stages of checking.
-- Allow `^` suffix for non-recursive imports #1480.
+- Allow `@norecurse` attribute for non-recursive imports #1480.
 - wasm32 / wasm64 targets are use-libc=no by default.
 
 ### Fixes